# Improve `String::sprintf` performance