--- a/core/string/ustring.cpp
+++ b/core/string/ustring.cpp
@@ -5678,7 +5678,7 @@ String String::sprintf(const Array &values, bool *error) const {
 					in_decimals = false;
 					break;
 				default:
-					formatted += chr(c);
+					formatted += c;
 			}
 		}
 	}