@kus04e4ek 2024-06-28
Makes sense to me! There are other instances of unnecessary `chr` though. Looks like `String::operator+=(char32_t p_char)` just didn't exist when this code was written and that was the reason why it wasn't used