@clayjohn 2024-07-24
Superseded by https://github.com/godotengine/godot/pull/94558 which includes a few other related changes. Thank you anyway as your work helps to further validate that these changes are necessary!